Vales of Anduin reorganized
To better represent the lore that describes the Vales of Anduin as a
"land that was for the most part deserted due to the dread of Mirkwood.
Only the Beornings, including Beorn himself, the Woodmen and Radagast at
Rhosgobel dared live in or near this land" and, speaking about Grimbeorn
"Like his father, he was the chieftain of the Beornings and he dwelt in
the same regions of the Vales of Anduin. During his time, the High Pass
and the Ford of Carrock were kept open and no Orc or Wolf dared to enter
his land". So, it's unlikely that Beorning controlled the entire
territory of the Anduin. I've reoorganized a bit these lands:
- Added some new Beorning characters from LOTRO
- Assigned some of the titles to these new holders
- Now Beornings holds the Beorncrofts and parts of the Carrock to the
High Pass, to better represent the lore, the other lands are wilderness.